Rename critical_pts module to critical
